[fix] create issue

+++ Se_Capstone_Backend/Capstone.Api/Module/Issues/Request/CreateIssueRequest.cs
@@ -9,7 +9,7 @@ namespace Capstone.Api.Module.Issues.Request
         public DateTime? StartDate { get; set; }
         public DateTime? DueDate { get; set; }
         public Priority? Priority { get; set; }
-        public int? EstimatedTime { get; set; }
+        public float? EstimatedTime { get; set; }
         public Guid? ParentIssueId { get; set; }
         public Guid? AssigneeId { get; set; }
         public Guid StatusId { get; set; }
